+++ docs/changelog.rst
@@ -4,7 +4,12 @@ Changelog
 
 * 2023-xx
 
+    * Fix small issues in sequences/BDSIMSequence
+    * Update pybdsim in pyproject.toml
+    * Fix in the documentation
+
 .. todolist::
+
 
 * 2022-1